fixes for types/linting
