+++ decs.d.ts
@@ -1,0 +1,7 @@
+import { SetupServerApi } from 'msw/lib/types/node';
+
+declare module 'vitest' {
+  export interface TestContext {
+    server?: SetupServerApi;
+  }
+}
